fix: make DEFAULT_CHARGE an int
Total charge is integral, and some external calculators (e.g. UMA's
FAIRChemCalculator) reject a float charge. Storing 0 instead of 0.0 fixes
those without affecting mlip models (which coerce internally).

+++ src/mlipaudit/benchmark.py
@@ -30,8 +30,9 @@ from mlipaudit.run_mode import RunMode
 RunModeAsString: TypeAlias = Literal["dev", "fast", "standard"]
 
 #: Default total charge applied to `atoms.info["charge"]` when a benchmark's
-#: input data does not specify one. Represents a neutral system.
-DEFAULT_CHARGE: float = 0.0
+#: input data does not specify one. Represents a neutral system. Kept as an int
+#: since total charge is integral and some calculators (e.g. UMA) reject floats.
+DEFAULT_CHARGE: int = 0
 
 #: Default spin multiplicity applied to `atoms.info["spin"]`. All benchmarks
 #: currently treat systems as closed-shell singlets unless overridden.
